\name{timetraveleR-package}
\alias{timetraveleR-package}
\alias{timetraveleR}
\docType{package}
\title{R/CRAN environment on VM using Vagrant}
\description{
  VirtualBox virtual machine manipulation using Vagrant to recreate a
  R/CRAN environment.
}

\details{
  CloneR consists of functions to install CRAN packages on a
  Vagrant/VirtualBox VM such that packages match the version that was on
  CRAN for a given date.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}

\keyword{virtualmachine}
\keyword{virtualbox}
\keyword{vagrant}
\keyword{cran}
